MFT-188

+++ i21Database/dbo/Stored Procedures/uspTFGenerateSF900.sql
@@ -32,7 +32,7 @@ DECLARE @SmrySummaryItemNumber NVARCHAR(MAX)
 DECLARE @SmrySummaryItemSequenceNumber NVARCHAR(MAX)
 DECLARE @SmrySummarySection NVARCHAR(MAX)
 DECLARE @SmrySummaryTaxType NVARCHAR(MAX)
-DECLARE @SmryTempTotal INT
+DECLARE @SmryTempTotal NUMERIC(18, 2)
 DECLARE @SmrySummaryItemsCount NVARCHAR(MAX) 
 DECLARE @SmryQuery NVARCHAR(MAX)
 --DETAIL VARIABLES
